refactor: :recycle: started first refactor
