+++ src/main.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import ReactDOM from "react-dom";
 /* import { BrowserRouter } from "react-router-dom"; */
 import { HashRouter } from "react-router-dom";
-import App from "./components/App";
+import App from "./App";
 
 ReactDOM.render(
   <React.StrictMode>
