+++ src/pages/Editor/TracksTab/index.tsx
@@ -1,3 +1,5 @@
+// todo - separar / variaveis não utilizadas
+
 import React, { useEffect, useState } from "react";
 import { Layout, Menu, Divider } from "antd";
 import {
